Switching core to DBI interface
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@1196 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/corelibs/U2Gui/src/util/CreateAnnotationWidgetController.h
@@ -51,7 +51,7 @@ public:
 
     GObjectReference        annotationObjectRef; // the object to be loaded
     QString                 newDocUrl;        // the URL of new document with annotation table to be created
-    int                     sequenceLen;        //length of target sequence for validation purposes
+    qint64                  sequenceLen;        //length of target sequence for validation purposes
 
     AnnotationTableObject*  getAnnotationObject() const;
 };
